moved page 4 file and stl added
